Allow one submission to override others in grading
This commit adds the `force_exercise_points` boolean field to the
`Submission` model. When this field is set to True, that submission's
points will determine the final score for the exercise, even if the
student has other submissions with better scores. This will allow
assistants to set the final points for the exercise when grading it
manually, without having to worry about the scores of other submissions.

Related to #831