Add button for closing the course sidebar
This commit adds a button to the course sidebar that closes it. When
closed, the sidebar can be reopened by clicking a bar that appears on
the left side of the screen. The state of the sidebar is stored in
localStorage, so it will remain closed upon navigation. This feature
improves the user experience especially on small displays and in views
that display a lot of information horizontally (e.g. the new assessment
view).

Related to #831