Delayed feedback (reveal rules & grading mode)
This commit implements new configuration features to A+ courses. Users can now
define RevealRules that determine when submission feedback and model solutions
are revealed to students. If submission feedback is not currently revealed,
students can't see the points they have received anywhere in the UI or API. In
addition, exercises can now be configured to use the last submission instead
of the best one in grading. This is done with a new grading_mode field. Both
the RevealRules and grading_mode can be configured in RST, and imported to A+.
There are various changes made to templates and point calculation code to
ensure that the new options are respected and unrevealed points are not
revealed to students.