Fix messages disappearing after first submission
There was a bug that caused exercise messages such as "Staff can submit
assignments without enrolling" and "Only your last submission will be graded"
to disappear forever after the student makes their first submission. This was
caused by chapter.js, which replaced the entire exercise container, including
the message area, when the last submission was loaded and displayed. This has
been fixed and now the messages are not removed. Additionally, the "Only your
last submission will be graded" message is now added in
`ExerciseView.submission_check` instead of `ExerciseView.get`, so it remains
visible after clicking the submit button.

Fixes #881