Allow passing normal_deadline to get_new_deadline
This commit adds the possibility to pass the normal deadline as an
argument to DeadlineRuleDeviation.get_new_deadline, if it is already
known by the caller. If the argument is provided, the method does not
query the normal deadline from the course module as it otherwise would.
This improves performance.

Related to #882