Reveal rule trigger DEADLINE_ALL optimization
This commit improves the performance of the DEADLINE_ALL trigger mode
of reveal rules. The exercise object is no longer fetched separately,
instead the deviations are queried directly with the exercise id found
from the cache. The found deviation is also cached within the
ExerciseRevealState object, so it won't have to be queried again if
RevealRule.is_revealed and RevealRule.get_reveal_time are called
separately. In addition, the known normal deadline is passed to
get_new_deadline, which avoids additional queries.

Related to #882