Add Django Debug Toolbar
The toolbar is a set panels that are added to the web page. They contain
useful debug information for A+ developers, for example, the SQL queries
that are executed in the database.

The toolbar can be activated in settings.py (or local_settings.py)
with the variable ENABLE_DJANGO_DEBUG_TOOLBAR.
By default, it is disabled. Note that the toolbar's pip package must
be also installed when the toolbar is enabled.

Closes #813

Co-authored-by: Jhosimar Aguacia <jhosimar.aguaciafisco@aalto.fi>