Add style for left/right joined buttons
This commit adds the styles aplus-button--left and aplus-button--right.
Two buttons using these styles appear to be joined together, i.e. the
corners between them are not rounded. This can be used to signify that
the function of the right-side button is related to the left-side one.
For example, the right-side button could be a dropdown button or a help
button.