Remove retry from AJAX submission code
This commit removes the retry logic from submitAjax in chapter.js, which
was originally added in pull request #310.

As explained in the pull request discussion, the retry logic was added
because submissions would sometimes fail to be saved in the database
(error message "database is locked"). This was due to threading issues
with SQLite, however, those are no longer relevant since A+ now uses
PostgreSQL in the Docker container.

Even though the retry logic originally checked for the database lock
error specifically, it was changed during the PR review process to be
used for all non-200 return codes. It was recently found that this would
lead to multiple submissions being created in case A+ returns an error,
because chapter.js would retry the submission 5 times.