Improve the inspect page messages for unofficial submissions
The messages used to include unofficial submissions as if they were
graded. For example, the message on a graded submission said that
another submission has more points even though the other submission
is an unofficial submission. Likewise, "there are newer submissions"
should not be shown if the newer submissions are unofficial.