Deviation list performace improvement
This commit significantly improves the deviation list view's performance
by using `select_related` and `prefetch_related`. There were multiple
database queries executed for each row in the list, which caused
performance issues with large numbers of deviations. Now, the number of
queries required for rendering the page is constant, and a list of 4000
deviations is rendered in a couple of seconds.