Submission list performace improvement
This commit significantly improves the performance of the "my
submissions" list by using `select_related` and `prefetch_related`.
There were multiple database queries executed for each row in the list,
which caused performance issues with large numbers of submissions.
Now, the number of queries required to render the list is constant.