Student information system support to course instances
Add support for linking cloned course instances to local
university student information system (SIS), such as SISU
in Aalto. This commit includes only the abstract
StudentInfoSystem base class, and no real implementation, which
is intended to be implemented as plugins by organisations using A+,
and thus not included in aplus repository.