Use staff CachedPoints to find the best submission
There was a bug in `InspectSubmitterView` and
`NextUnassessedSubmitterView` where the student version of
`CachedPoints` would be used to find the best submission to redirect to.
This would always be the last one if delayed feedback is enabled, even
though course staff (the users of these views) is supposed to be able to
know which submission is the best one.