Allow passing an is_revealed value to points badge
There was an issue where reveal rules were calculated for each
submission in the submission list separately, even though the result
would be the same for each submission. This was fixed by allowing an
`is_revealed` argument to be passed
to the `points_badge` and `points_progress`
template tags, when it is known.