Fix CachedPoints duplication
`CachedPoints` was duplicated in the delayed feedback update because it
was the easiest way to store revealed and unrevealed versions of the
data simultaneously. There were two versions of the cache for each
student, one where parameter `is_staff` was `True` and one where it was
`False`. This commit fixes the duplication by storing both values in the
same cache, prefixed with `_staff_` or `_student_`. In the
`CachedPoints` initializer, one set of these keys is exposed while the
other set is removed.