View updates for submission drafts
This commit adds a new `SubmissionDraftView` for updating submission
drafts. Additionally, `ExerciseView` has been updated so that it can
return a submission page or a draft page, if requested with URL
parameters. This makes it easier to load the appropriate page in the UI
without multiple HTTP requests. In addition, existing view code has been
updated so that submission drafts are deactivated when the user makes a
submission or changes groups.

Related to #837