add publish functionality to configure_content
The service can return a publish_url field which will be called after
the configuration is succesful: no publish happens if it fails.
The publish url may also fail, in which case the configure is rolled
back as the assumption is that the update failed. The publish url
should return a JSON dict with "errors" containing errors and "success"
telling whether the publish was succesful (defaults to True). For non-dict
response, publish is assumed to have failed.