New AplusAutoSave jQuery plugin
This commit adds the new jQuery plugin `AplusAutoSave`, which is used in
exercises that have the `data-aplus-autosave` attribute set. This causes
a submission draft to be saved periodically while the user types their
answer, and the draft is restored when returning to the page later. In
addition, chapter.js now makes use of `ExerciseView`'s new URL
parameters to load the appropriate page (draft/last submission/blank
form) without multiple HTTP requests. Because of this,
`loadLastSubmission` is no longer called for quiz exercises.

Related to #837