Display deviations in student UI
This commit adds information to the student UI about the deviations
granted to the student. Max submissions deviations are now counted in
places that display a submission counter in the format X/Y where X is
the current number of submissions and Y is the maximum number of
submissions. Deadline deviations are now displayed in the "exercise
info" box on the standalone exercise and submission pages. They are also
displayed in the exercise frame header.

Related to #908