Order submissions in CachedPoints
This commit changes the submission list in CachedPoints to be ordered by
exercise and submission time. Because of this change, the final and last
submission of each exercise no longer has to be stored in a dict. In
addition, small optimizations can be done, such as fetching `entry` in the
submission loop only once per exercise. Reveal rules are now evaluated
after a single exercise has been handled, instead of after all exercises,
and because of this, deviations are now handled before submissions.

Related to #887