Fix crashes in the aggregate data API when querying exercises
The aggregatedata API endpoint always crashed when exercises or
chapters were queried (`?filter=1.1` or `?exercise_id=1`),
for example, `/api/v2/courses/1/aggregatedata/?exercise_id=1`.

The crash was a KeyError on the line
`exercise_max[num] += e['max_points']`
in exercise/api/csv/aggregate_sheet.py (the second if branch
inside the `exercises` for loop).
The `num` variable is None because the first if branch is not
entered in the loop. The None value causes KeyError when it is
used as a dictionary key. Normally, the `exercises` list contains
module and exercise objects. The module object enters the first
if branch and initializes variables like `num` for the following
exercise objects. However, when a single exercise is queried,
the `exercises` list contains only the single exercise object.
In that case, the variables were not initialized and the code
crashed.

This commit changes the `search_entries` method of CachedContent.
Its `entry` return value is only used in the aggregatedata API view.
In this commit, instead of returning None for `entry`, the cache
object for the queried exercise, chapter or module is returned.
This changes the `number` parameter in the `aggregate_sheet`
function so that the exercises for loop enters the first if branch.
The `num` variable is initialized correctly and everything works.
The condition of the second if branch is changed so that it does
not accept the root object even if the root is an exercise or
chapter. (Previously, the code worked only when the root was a
module.) Note that chapters have `type: "exercise"` in the cache.