Fix a crash in exercise.load due to missing parameters
Some BaseExercise subclasses did not include all possible
parameters in the definition of the `load()` method.
This caused a crash when the method was called with those
parameters. The method `Submission.load()` in particular supplies
the `ordinal` argument to `exercise.load()`.
The `Submission.load()` method was recently added in
the delayed feedback feature (pull request #880).

For example, Rubyric exercises using delayed feedback crashed after
submitting. Luckily, the submissions were saved to the database
before the code crashed in `Submission.load`.