Move table ordering to AplusTableFilter
Table ordering was originally implemented in participants.js, so it was
specific to the participants page. To facilitate using it elsewhere, it
has been refactored into the AplusTableFilter jQuery plugin, which
already implemented table filtering. Ordering can be enabled with the
"ordered-table" class( much like the old "filtered-table" class) and
disabled for individual columns with the data-order-disable attribute.

Ordering has been enabled for all tables that already had filtering.