Add grouping to AplusTableFilter
This commit adds grouping functionality to the AplusTableFilter jQuery
plugin. Parent rows of a group should be marked with the attribute
data-parent-group="<group-id>" and their child rows with the attribute
data-parent-child="<group-id>", and the table should have the
grouped-table class. A column with expand/collapse buttons is added to the
left side of the table.