Fix 404 when clicking "start manual assessment"
The "start manual assessment" button, as well as the "inspect" button in
the submission list when grouped by submitter, would cause a 404 error
if none of the chosen user's submissions were marked as
`best_submission` in the points cache. This commit fixes the issue by
selecting the user's last submission regardless of its status, when
there is no "best" submission.