Exclude WAITING submissions from all results table
In a previous commit (87dae322fd9777f707db2d0279ccfd1258bf0587),
UserExerciseSummary was changed to include submissions in states WAITING
and INITIALIZED, to match the behaviour of the all results table (which
uses CourseResultsDataViewSet). It was later decided that this should be
the other way around: the all results table should ignore submissions
in the aforementioned states, like UserExerciseSummary and CachedPoints
did. This commit reverts the previous UserExerciseSummary change and
changes CourseResultsDataViewSet to ignore submissions in these states.
test_annotate_submitter_points was also expanded for more comprehensive
testing of annotate_submitter_points.