Synchronized checkboxes for table.js
This commit adds a new synchronized checkboxes feature to table.js. This
means a checkbox placed on a parent row of a grouped table will react as
the checkboxes on the child rows are checked/unchecked. If the parent
checkbox is checked/unchecked by the user, all child checkboxes react to
it.

Fixes #896
Fixes #902