Use type variable in JWT decorator
The `register_jwt_accessible_class` was typed using `Type[Any]`. Because
of this, IDE type checking and autocompletion was broken for all classes
that the decorator was applied to. This commit changes the decorator to
use type variables to signify that the input and output types of the
decorator are the same, which restores the IDE experience.