Do not exclude submitters with no ready submissions in the list
The submitters list used to exclude submitters who have only
unofficial or erroneous submissions. However, they were included
in the counter (the number of submitters at the top of
the submissions list). The teacher could be interested in submitters
who have only erroneous or unofficial submissions, thus they should
be included in the submitters list.