Add PENDING enrollment status for SIS enrollments
The PENDING enrollment status is used for courses with enrollment
questionnaires, when students are enrolled from SIS. PENDING students
cannot start the course tasks before completing the enrollment
questionnaire.

Use the PENDING status also for non-SIS courses when the student
opens the enrollment questionnaire. If they do not submit
the questionnaire successfully, the PENDING enrollment will stay
in A+.

PENDING enrollment may be activated manually to the ACTIVE status
(in the `enroll_student` function).

Fixes #981.