Add late and unofficial submission approval feature
Course staff may approve late and/or unofficial submissions so that
any point penalties are removed from those submissions and
they become normal, graded submissions (with the "ready" status).

Single submissions may be approved or all submissions from
one student in one assignment or one whole course module.
The whole assignment or module approval may also be set to target
only late submissions or only unofficial submissions.

The user interface for this feature is visible in the inspect
submission view when inspecting a late or unofficial submission.

Part of #892.
"The first pull request" in the comment:
https://github.com/apluslms/a-plus/issues/892#issuecomment-1048656166

This work started in Muhammad's pull request #991. Markku has
refactored the code, improved the database queries and the details
overall.

Co-authored-by: Muhammad Wahjoe <faiz00.muhammad@gmail.com>