WIP Approve late/unofficial submissions when adding deviations
TODO:
* +1 max submissions deviation button in the inspect submission view:
  should this also check if any submission should be approved
  immediately?
* test all the changes
* deviations/viewbase.py approve_unofficial_submissions():
  implementation unfinished
* clean up the code
* success messages to the user in the user interface?

When deviations are added, the teacher may choose to also immediately
approve such existing submissions that are covered by the deviation.
For example, if the submission was late, but with the new personal
deadline it is not late, then it would be approved.

Fixes #892.
"The second pull request" in the comment:
https://github.com/apluslms/a-plus/issues/892#issuecomment-1048656166

This work started in Muhammad's pull request #991. Markku has
refactored the code, improved the database queries and the details
overall.

Co-authored-by: Muhammad Wahjoe <faiz00.muhammad@gmail.com>