Add best results API endpoint that ignores grading mode LAST
Add a separate best resultsdata API endpoint which ignores
the last-only grading rule in exercises. This API endpoint is used
by default in the All results view and the previous, normal
resultsdata endpoint can be enabled with a checkbox in the All results
view.

This is a temporary, hacky fix for #918.
The database query for handling exercises using the grading mode
LAST performs too slowly. The code for it is in
`SubmissionQuerySet.annotate_submitter_points()`.