Fix bugs and coding style issues in the download buttons
Commit 9e3b5b9d included some bugs and coding style issues
that are fixed in this commit.

The inspect view had two download buttons for submitted files
because assessment.js had its own, older implementation for
the download button.

The HTML class `pre-content` was unnecessary and it made
the JavaScript function `highlightCode()` less reusable since
callers of the function would have to include the HTML class in
the `pre` element.