Show the "hidden from students" label on the course archive page
The course archive page and the course instance list
(under one course) did not show the "hidden from students" label
on hidden course instances. This commit fixes that.

The template `_course_cards.html` now always shows the hidden label
when the course instance is hidden. There is no use case when
the label should not be shown on a hidden course.

Fixes #727