Add more edit links to the notification admin view
The notification admin view used to add links to
the "edit notification" page only in the "subject" column.
However, if the subject is empty, there is no edit link at all,
which makes the admin's life frustrating. This commit adds
the link to other columns too so that the link is always available.