Add support for MathJax 3.x JavaScript
This commit does not break the support for MathJax 2.x. Rather than
unconditionally using the 2.x API (Hub.Queue([])), we instead check
the version of MathJax and either use the old Hub.Queue implementation
or the new promise based implementation of 3.x
(MathJax.typesetPromise()). This conditional typesetting is defined
as a jQuery plugin in assets/js/aplus.js.

Fixes #871