Improve database performance in get_max_deviations() / points API
This method is used in `CachedPoints`, which is used in the points API.
The points API has been performing very slowly due to this database
query that fetches the (deadline or submission attempt) deviations
for the student. The OR operator in the old query made it slow.
The new version does not use the OR operator and it performs over
10 times faster.

Fixes #1087