Connect form labels correctly to the input in search-select widgets
In an HTML form, clicking a label should move the focus on
the corresponding input element.
This commit fixes that behavior for search-select form widgets.
The widget is used, e.g., in the assistant field of
the edit course form.

The widget id attribute given in the parameters of
the `get_context()` method is set on the input HTML element
while the hidden select element is given a new, different id
by adding `_select` to the end of the id.

Fixes #934
Closes #1071