Fix the relative URL used in the last course instance redirect
The old implementation used the short URL key of the course instance
in the HTTP Location header. This caused an additional HTTP redirect
response due to the trailing slashes `/` in URLs.
Use the full URL path instead in the Location header.
The method `get_display_url()` from the `CourseInstance`'s
superclass `UrlMixin` is intended for these use cases.