Fix the id field used in the field-values filter in the user list APIs
The user and student API viewsets use a UserProfile queryset, but
the id values given in the URLs are User.id values,
not UserProfile.id values. The old buggy code was using
the given id as a UserProfile.id, which returns wrong users when
the database has different ids for corresponding User and UserProfile.

Fixes #1091