Add granter to the deviation group header in the list view
In the deviation list view, when deviations are grouped together,
the group header did not show the granter and grant time columns
even if they were identical for each record in the group.
This commit fixes it.

If the granter is not the same for each deviation in the group,
the deviations are still grouped but the group header does not
show any granter.

Fixes #1046