Add LTI Tool v1.3 feature
A+ may be used as an external tool from LTI platforms such as
Moodle.

There are some unfinished parts left in this feature, for example,
the LTI grading service is missing for Acos server assignments
and other assignment services that create new, graded submissions
by sending the submission data via the assignment service in the end
of the student's workflow.

The discussion in the pull request contains more items to be fixed
later.
https://github.com/apluslms/a-plus/pull/1104

Mikko and Eerik implemented small parts of the feature.
Markus did most of the work in this commit.

Co-authored-by: Murhu Mikko <mikko.murhu@aalto.fi>
Co-authored-by: Saksi Eerik <saksi.eerik@gmail.com>