Fix submission feedback polling in the full-view exercise page
The A+ LTI Tool v1.3 (#1104) implementation broke the polling of
the submission feedback in full-view exercise pages.
The polling was also broken in the LTI exercise full-view pages.

Remove the URL `lti-submission-poll` because it used the same view
as the normal `submission-poll` URL. The template
`_lti_exercise_wait.html` is not needed since it was a copy of
`_exercise_wait.html` with a different poll URL.

Poll.js is changed so that the final target URL can be configured
through HTML data attributes. Previously, poll.js had a hardcoded
assumption that the final target URL could be derived from
the poll URL by dropping the word "poll" from the end of
the poll URL. That was not flexible enough for LTI Tool URLs that
start with the "/lti" prefix.

The template block `exercise_wait` is now unnecessary.
It was added in #1104. Adding the block was a mistake originally
since the parent templates did not have such a block and
it basically did not render anything. The block `exercise_wait`
was not part of any existing block.

The template `lti_submission.html` should not override
the block `exerciseinfo`. It was copied from `submission.html` and
the submission info page was thus rendered twice in
the LTI submission page.

Fixes #1114
Closes #1162