Move submit modal dialog error message from progress bar to alert box
In the submit modal dialog in chapters, the error message
was displayed inside a progress bar. That was not a good idea
since it was hard to read and the latter half of the message
disappeared since it did not fit into the progress bar. The
error message is now moved to a separate alert box above
the progress bar.