Add support for nested cache items
It uses cache.get_/set_many. This is slow on pymemcache, so pylibmc
should be used instead.

A cached object can contain another cached object. The nested
object is stored as a 'proxy' object that only has the cache_key and
params stored. This proxy can then be resolved to the actual object by
fetching the data from the cache.

The nested cached objects are fetched lazily on-demand unless
they are fetched earlier with .populate_children() or
prefetch_children = True.

The proxy fetching system is smart in that it will only fetch an object
once even if there are multiple proxies that resolve to that same
object. Of note is that the proxies that resolve to the same object will
also share the same data: modifying one of them will modify all of them.