Increase the chapter.js submit timeout to 20 seconds
Timeout was added to the AJAX submissions in chapter.js in
the pull request #1185. However, 10 seconds was too short compared
to the retry loop when A+ is not able to immediately connect to
the grader. Increase the timeout to 20 seconds so that A+ can
finish the connection attempts to the grader before chapter.js
causes a timeout for the submission attempt.

Pull request comment:
https://github.com/apluslms/a-plus/pull/1185#discussion_r1248016673
The retry loop in the code:
https://github.com/apluslms/a-plus/blob/88f2dc91667eed12a275427585643b6c787dfa1c/lib/remote_page.py#L60-L62