Add consecutive_order, parent_name and status to exercise API
Add a consecutive numbering to the API to indicate the order of learning
objects (chapters and exercises). This is calculated dynamically.

Add a field to the API "parent_name", which contains the name of the
parent, i.e., the chapter or exercise the exercise is embedded in (if it
is embedded).
The name of the chapter can provide more understandable context than
just the exercise path or the title of the exercise (e.g. "Feedback").

The consecutive_order field is used to fix Jutut issues #43 and #44.
The parent_name field is used to fix Jutut issue #46.
The status can be used to identify enrollment questionnaires, needed
for fixing Jutut issue #31.