Save points cache objects individually
Points cache objects are cached separately and are built on top
of the content cache objects.

- Added invalidate_time to each object, so that the data is invalidated
when the points are supposed to be revealed.
- unconfirmed field was replaced with a property that check the parent's
children_unconfirmed field. This is because determining the value
of unconfirmed during generation requires looking at the sibling
learning objects with confirm_the_level = True and this causes
a cyclical dependency in the generation of exercises.