Add dependency support to cache objects
- Makes invalidation much easier to deal with and more reliable. As far
as I can tell, this basically quarantees the validity of the cache data
(as long as the database wasn't updated without the signals being run,
which shouldn't happen)
- The bad side is that the whole dependency tree must be loaded whenever
a cache object is loaded. This makes the loading slower but doesn't seem
to be so slow that it's a problem.
- _generate_data can return dependencies that are checked for validity
when the object is loaded from cache. The object is regenerated if the
validity fails. Note that if the dependency is newer, the object is
considered invalid as the dependency has changed after the object's
generation.
- Generally speaking, INVALIDATORS should be used if the model object
is accessed directly inside _generate_data and dependencies are used
when the data is accessed through a cache object or the object stores
references to other cache objects. Note that dependency is not needed if
the reference can only change when an invalidator fires. For example,
the module reference (i.e. the module id) of an exercise can only change
if the exercise model changes because the module id is stored in the
exercise model. This means that the module does not need to be added to
the dependencies (as long as the exercise cache is invalidated through
the INVALIDATORS whenever the exercise model changes).
- is_valid was removed as it would require building each dependency
object to check if an object is valid. Instead, the expiry time is
stored alongside the generated_on on value to allow invalidating using
an expiry time without the need to build each dependency.