Set programming language manually for highlight.js
Highlight.js may automatically detect the programming language
whose syntax is highlighted, but the latest version of highlight.js
fails to detect the language correctly more often than expected.

This fix sets the language manually based on the URL file extension.
If the URL does not end with a file extension or highlight.js does not
know the file extension, the language is detected automatically.

Fixes #1221