Fix sticky alerts that hide information when using anchor links
Course pages may show a sticky alert that the course instance has
ended. The alert may hide information when anchor links are used
since the alert goes on top of the anchor target like a heading.
This commit adds padding to the target so that only empty space
is really covered by the sticky alert at the start before
the user starts to scroll the page.

Fixes #1207