Chapter may be set as a model answer to course modules
Add a new feature that enables controlling chapter visibility to
students based on the completion of a course module.
There are multiple alternative rules for revealing the chapter
since the feature uses the same "reveal rules" feature that has
previously been developed for the visibility of
assignment model solutions and submission feedback.

Fixes #1098