Fix deleting student tags (since Django 4.0)
Django 4.0 changed `DeleteView` so that A+ `UserTagDeleteView`
started to fail due to form validation that never passes.
The student tag was never deleted and the web page seemed to be
stuck in the delete confirmation.

In Django 4.0, `DeleteView` uses the attribute `form_class`.
`UserTagDeleteView` already had `form_class` via inheritance from
`UserTagMixin`.

The validation in `UserTagForm` is not suitable for
`UserTagDeleteView`. Unwanted validation is removed from
`UserTagDeleteView` by setting `form_class` to
the empty `Form` base class.

https://docs.djangoproject.com/en/4.2/releases/4.0/#generic-views
https://docs.djangoproject.com/en/4.2/releases/4.0/#deleteview-changes

> DeleteView now uses FormMixin, allowing you to provide a Form
> subclass, with a checkbox for example, to confirm deletion.
> In addition, this allows DeleteView to function with
> django.contrib.messages.views.SuccessMessageMixin.
> In accordance with FormMixin, object deletion for POST requests
> is handled in form_valid(). Custom delete logic in delete()
> handlers should be moved to form_valid(), or
> a shared helper method, as needed.

Fixes #1261