Fix crash in CachedPoints with GraderUser from commit 0e74906b
Commit 0e74906b156b9f3da07ac198760d1133e5038b2a (#1273) introduced
a crash when `GraderUser` accesses any course page or API.
The commit added a `CachedPoints` instance variable to
`CourseInstanceBaseMixin`, but `CachedPoints` crashes when it is
initialized with the `GraderUser`. `CachedPoints` does not crash
with `AnonymousUser`, thus it is used instead of `GraderUser`
to work around the crash.

In the master branch, the commit
0c0de05a2e2fa9b14d2247088c5ad87bc9592884 is similar to 0e74906b,
but there is no crash due to the cache changes in #1231.