Fix bugs in the LTI Tool views' course content links
The LTI course view and module view had wrong links to
the course contents: the links pointed to the normal views
instead of the LTI versions of the views.

There was a crash in the LTI exercise view: `upper_name` and
`upper_link` variables were used when they were not defined at all.

Related to #1239 and #1290