Restrict access to hidden courses: LTI session targets one course
After using the A+ LTI Tool, e.g., from a Moodle LTI activity,
students were able to access any hidden course in A+.
`CourseVisiblePermissionBase` used to skip
the `course_instance.visible_to_students` setting if the user had
an LTI launch id in the A+ session. The session data remains even
after the student stops using the LTI Platform (like Moodle) and
accesses A+ directly.

I was able verify that the student was able to access other hidden
courses in A+ and also submit to the assignments. The Moodle LTI
activity needs to be launched into a new window since the session
cookies seem to behave differently in iframes. After the student
has launched the LTI activity in Moodle, the student may open
the normal A+ site in another browser window. The student is
already logged in there. Then, the student may manually type
the A+ URL of some hidden course. The student is able to browse
the hidden course normally. However, the student is not able
to enroll in the hidden course, so they should be already enrolled
or the course should have the setting "view content to" in
a less strict value than "enrolled students".

The issue is fixed by checking that when the student is accessing
a hidden course, it matches the course set in the LTI launch data.
Thus, students may only access a hidden course if they have access
to the course via LTI, e.g., from Moodle.
Note that chapter views in the LTI Tool access embedded exercises
through the normal exercise view.
Thus, the function `has_lti_access_to_course()` looks a bit complex
since it needs to be able to check the LTI session even if
the current view is not any LTI view.

Related to #1240 and #1290