LTI Tool v1.3: add function for checking the course in the launch
The function `has_lti_access_to_course()` is used to check if
the user's LTI launch targeted the given course instance.

The function looks a bit complex since it needs to be able to
check the LTI session even if the current view is not any LTI view.

In the v1.20_stable branch,
commit 8e6d873543212e8ddec24d0d5360e53befb19bc2 added the function
`has_lti_access_to_course`.