Hide module model solution chapters from unauthenticated users
When a course instance is public (visible without authentication),
unauthenticated users used to be able to view chapters that should
be hidden based on the course module model solution reveal rule.
(This also required that the model solution chapter does not have
the audience setting that requires users to authenticate.)
This issue is fixed by checking the course module model solution
reveal rules in `CachedPoints` even when the user is unauthenticated.
In practice, the code only had to be changed so that
the module list variable is not set to the empty list when
the user is unauthenticated. One function definition also had to
be moved out from an if branch.

The scenario described in the issue #1247 was effectively already
fixed in the previous commit that fixed the permission error
messages for permissions that are combined from other permissions
with the `OR` operator. This commit focuses on the visibility bug
described above.

Fixes #1247