Fix build log URL updating slowly
Build log URL did not update immediately when creating
a new course instance, instead it used the build log URL
from the previous course instance until the build was done.

It also did not update immediately when changing the Git Manager
key in the Git Manager tab in A+ course settings.

This caused an error when attempting to view the build log in A+
while the course was still being built.

Fixes #1339