Fix issues with points badge colors
The logic for counting exercises as passed, and thus coloring the points badges, was previously changed as such that unpassed children caused the parent to count as not passed. This has now been changed to work similarly as before so that exercises / chapters only consider themselves unpassed if their own points-to-pass has been defined and the points threshold has not been reached --> status of children won't be considered.

Also fixed an issue where exercises with 0 submissions were counted as unpassed even if they didn't have points-to-pass defined.